backend/requirements.txt
python-dotenv==1.0.0
google-generativeai==0.3.1
starlette==0.36.3
uvicorn==0.27.1